cdk/requirements.txt
aws-cdk-lib==2.32.1
constructs>=10.0.0,<11.0.0
typeguard==2.13.3
